requirements.txt
alembic==1.13.2
blinker==1.8.2
chardet==5.2.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
